present: Remove superfluous set abort flip function in window mode
The function is never called from present_screen.c in contrast to the behavior
in screen mode.

In present_wnmd.c we can simply remove the function which does an unnecessary
check of the property before setting it and directly set the property at the
two locations the function was called previously.

Signed-off-by: Roman Gilg <subdiff@gmail.com>